python-fastapi-simple-api/pyproject.toml
[tool.poetry]
name = "python-fastapi-simple-api"
version = "0.1.0"
description = ""
authors = ["Yuri Matheus <yurimatheusb@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
opentelemetry-instrumentation-fastapi = "^0.46b0"

[tool.poetry.group.test.dependencies]
pytest-randomly = "^3.15.0"
pytest = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
